third_party/blink/web_tests/external/wpt/css/css-text/i18n/unknown-lang/css-text-line-break-cj-strict-expected.txt

This is a testharness.js-based test.
Found 51 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] 3041  HIRAGANA LETTER SMALL A may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 3043  HIRAGANA LETTER SMALL I may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 3045  HIRAGANA LETTER SMALL U may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 3047  HIRAGANA LETTER SMALL E may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 3049  HIRAGANA LETTER SMALL O may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 3063  HIRAGANA LETTER SMALL TU may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 3083  HIRAGANA LETTER SMALL YA may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 3085  HIRAGANA LETTER SMALL YU may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 3087  HIRAGANA LETTER SMALL YO may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 308E  HIRAGANA LETTER SMALL WA may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 3095  HIRAGANA LETTER SMALL KA may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 3096  HIRAGANA LETTER SMALL KE may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 30A1  KATAKANA LETTER SMALL A may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 30A3  KATAKANA LETTER SMALL I may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 30A5  KATAKANA LETTER SMALL U may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 30A7  KATAKANA LETTER SMALL E may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 30A9  KATAKANA LETTER SMALL O may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 30C3  KATAKANA LETTER SMALL TU may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 30E3  KATAKANA LETTER SMALL YA may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 30E5  KATAKANA LETTER SMALL YU may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 30E7  KATAKANA LETTER SMALL YO may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 30EE  KATAKANA LETTER SMALL WA may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 30F5  KATAKANA LETTER SMALL KA may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 30F6  KATAKANA LETTER SMALL KE may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 30FC  KATAKANA-HIRAGANA PROLONGED SOUND MARK may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31F0  KATAKANA LETTER SMALL KU may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31F1  KATAKANA LETTER SMALL SI may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31F2  KATAKANA LETTER SMALL SU may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31F3  KATAKANA LETTER SMALL TO may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31F4  KATAKANA LETTER SMALL NU may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31F5  KATAKANA LETTER SMALL HA may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31F6  KATAKANA LETTER SMALL HI may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31F7  KATAKANA LETTER SMALL HU may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31F8  KATAKANA LETTER SMALL HE may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31F9  KATAKANA LETTER SMALL HO may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31FA  KATAKANA LETTER SMALL MU may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31FB  KATAKANA LETTER SMALL RA may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31FC  KATAKANA LETTER SMALL RI may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31FD  KATAKANA LETTER SMALL RU may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31FE  KATAKANA LETTER SMALL RE may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] 31FF  KATAKANA LETTER SMALL RO may NOT appear at line start if strict
  assert_approx_equals: expected 99 +/- 1 but got 69
[FAIL] FF67  HALFWIDTH KATAKANA LETTER SMALL A may NOT appear at line start if strict
  assert_approx_equals: expected 84 +/- 1 but got 54
[FAIL] FF68  HALFWIDTH KATAKANA LETTER SMALL I may NOT appear at line start if strict
  assert_approx_equals: expected 84 +/- 1 but got 54
[FAIL] FF69  HALFWIDTH KATAKANA LETTER SMALL U may NOT appear at line start if strict
  assert_approx_equals: expected 84 +/- 1 but got 54
[FAIL] FF6A  HALFWIDTH KATAKANA LETTER SMALL E may NOT appear at line start if strict
  assert_approx_equals: expected 84 +/- 1 but got 54
[FAIL] FF6B  HALFWIDTH KATAKANA LETTER SMALL O may NOT appear at line start if strict
  assert_approx_equals: expected 84 +/- 1 but got 54
[FAIL] FF6C  HALFWIDTH KATAKANA LETTER SMALL YA may NOT appear at line start if strict
  assert_approx_equals: expected 84 +/- 1 but got 54
[FAIL] FF6D  HALFWIDTH KATAKANA LETTER SMALL YU may NOT appear at line start if strict
  assert_approx_equals: expected 84 +/- 1 but got 54
[FAIL] FF6E  HALFWIDTH KATAKANA LETTER SMALL YO may NOT appear at line start if strict
  assert_approx_equals: expected 84 +/- 1 but got 54
[FAIL] FF6F  HALFWIDTH KATAKANA LETTER SMALL TU may NOT appear at line start if strict
  assert_approx_equals: expected 84 +/- 1 but got 54
[FAIL] FF70  HALFWIDTH KATAKANA-HIRAGANA PROLONGED SOUND MARK may NOT appear at line start if strict
  assert_approx_equals: expected 84 +/- 1 but got 54
Harness: the test ran to completion.
